requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
Pillow==9.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-vlc==3.0.16120
pytweening==1.0.4
Werkzeug==2.2.1
